% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/buscar_num.R
\name{buscar_num}
\alias{buscar_num}
\title{Search function}
\usage{
buscar_num(datos, x)
}
\arguments{
\item{datos,}{vector; x, searched number}
}
\description{
This function allows you to find the first ocurrence of a number in a vector
}
\examples{
buscar_num(-1000:1000,5) 1006 ; buscar_num(-1000:1000,2000) NA
buscar_num()
}
\keyword{buscar}
\keyword{num}

