% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/suma.dos.numeros.R
\name{suma.dos.numeros}
\alias{suma.dos.numeros}
\title{Sum two numbers}
\usage{
suma.dos.numeros(a, b)
}
\arguments{
\item{a}{number, b number}
}
\description{
This function allows you to sum two numbers
}
\examples{
suma.dos.numeros()
}
\keyword{suma.dos.numeros}

